# Conflicts:
#	src/main/java/ch/uzh/ifi/hase/soprafs23/service/VoteService.java

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/controller/VoteController.java
@@ -23,7 +23,7 @@ public class VoteController {
     public void saveVote(@PathVariable("gamePin") int gamePin,
                          @PathVariable("categoryName") String categoryName,
                          @RequestHeader("Authorization") String userToken,
-                         @RequestBody Map<Long, String> votings){
+                         @RequestBody Map<Integer, String> votings){
 
         voteService.saveVote(gamePin, categoryName, userToken, votings);
