+++ src/main/java/ch/uzh/ifi/hase/soprafs23/rest/dto/user/UserGetDTO.java
@@ -7,7 +7,7 @@ import java.util.Date;
 
 public class UserGetDTO {
 
-    private Long id;
+    private int id;
     private String username;
     private UserStatus status;
     private LocalDate creationDate;
@@ -22,9 +22,9 @@ public class UserGetDTO {
         this.quote = quote;
     }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
 
     public String getUsername() { return username; }
